+++ frontend/src/pages/AppManagement.vue
@@ -36,8 +36,9 @@
                             Here you can generate a new application secret.
                             This will <em>invalidate</em> the existing secret.<br>
                             See also
-                            <a href="https://github.com/bravecollective/brvneucore/blob/master/doc/documentation.md#authentication-of-third-party-applications"
-                                target="_blank">Authentication of third-party applications</a>.
+                            <a v-cloak target="_blank" :href="settings.customization_github +
+                                    '/blob/master/doc/documentation.md#authentication-of-third-party-applications'">
+                                Authentication of third-party applications</a>.
                         </p>
                         <p>
                             <button type="button" class="btn btn-warning" v-on:click="generateSecret()">
@@ -98,6 +99,7 @@ module.exports = {
         route: Array,
         swagger: Object,
         player: [null, Object],
+        settings: Object,
     },
 
     data: function() {